Define default `#to_s` methods for models
Why? Administrate uses `#to_s` to display objects throughout the admin
dashboard. Ruby's default `#to_s` method depends on the object id, which
changes with each request. That makes it impossible to keep track of
objects in the app.

This change adds `#to_s` methods that are consistent between different
requests, making it possible to differentiate between objects in the
dashboard.